Allow choosing the QR code error correction
As per the specification [1], the error correction should be 25%, which
is `ERROR_CORRECT_Q`. This allows regenerating exactly the same QR code
as the official one, and the TousAntiCovid application doesn't sometimes
complain anymore while importing it, saying that the format is invalid.
This commit allows choosing the wanted error correction, but keeps the
default to "H", in a conscious attempt at breaking the rules (and also
because it allows more fun in the picture choice).

[1]: sec 5.2.2, https://eur-lex.europa.eu/legal-content/FR/TXT/PDF/?uri=CELEX:32021D1073